requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.75.0
fastapi-mail==1.0.6
greenlet==1.1.2
h11==0.13.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==4.1.4
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.15.0
watchgod==0.8
websockets==10.2
wrapt==1.14.0
